requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.30
 
psycopg2-binary==2.9.9
 
gspread==6.1.2
gspread-dataframe==3.3.1
google-auth==2.34.0
 
python-dotenv==1.0.1